Fully functional MVP

+++ src/transaction.rs
@@ -1,18 +1,16 @@
 use serde::{Deserialize, Serialize};
 
-
-
 #[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Transaction{
+pub struct Transaction {
     sender: String,
     recipient: String,
-    amount: f32,
+    amount: f64,
     timestamp: i64,
 }
 
-impl  Transaction {
-    pub fn new(sender: String, recipient: String, amount: f32) -> Self{
-        Transaction{
+impl Transaction {
+    pub fn new(sender: String, recipient: String, amount: f64) -> Self {
+        Transaction {
             sender,
             recipient,
             amount,
